fix: depositMetadata.custom_accrual_rate as u8, completed test suite2

+++ programs/restaking/src/events/user_deposited_token_to_fund.rs
@@ -15,7 +15,7 @@ pub struct UserDepositedTokenToFund {
     pub supported_token_user_account: Pubkey,
 
     pub wallet_provider: Option<String>,
-    pub contribution_accrual_rate: Option<f32>,
+    pub contribution_accrual_rate: Option<u8>, // 100 is 1.0
     pub deposited_supported_token_amount: u64,
     pub minted_receipt_token_amount: u64,
 }
